Backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
certifi==2024.7.4
click==8.1.7
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.1
greenlet>=3.1.0
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy>=2.1.0
openai==1.42.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pgvector==0.3.2
pinecone[grpc]>=3.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary>=2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.30.6
wcwidth==0.2.13
zipp==3.20.0
redis==5.0.1
qdrant-client>=1.11.0,<2.0.0
